+++ src/config-type.ts
@@ -1,6 +1,4 @@
-import compression from "compression";
 import { Express, Request, RequestHandler } from "express";
-import fileUpload from "express-fileupload";
 import { ServerOptions } from "node:https";
 import { Logger } from "winston";
 import { AbstractEndpoint } from "./endpoint";
@@ -67,23 +65,6 @@ export interface CommonConfig<TAG extends string = string> {
   tags?: TagsConfig<TAG>;
 }
 
-type UploadOptions = Pick<
-  fileUpload.Options,
-  | "createParentPath"
-  | "uriDecodeFileNames"
-  | "safeFileNames"
-  | "preserveExtension"
-  | "useTempFiles"
-  | "tempFileDir"
-  | "debug"
-  | "uploadTimeout"
->;
-
-type CompressionOptions = Pick<
-  compression.CompressionOptions,
-  "threshold" | "level" | "strategy" | "chunkSize" | "memLevel"
->;
-
 export interface ServerConfig<TAG extends string = string>
   extends CommonConfig<TAG> {
   /** @desc Server configuration. */
@@ -97,15 +78,19 @@ export interface ServerConfig<TAG extends string = string>
      * */
     jsonParser?: RequestHandler;
     /**
-     * @desc Enable or configure uploads handling.
-     * @default false
+     * @desc Enable and configure file uploads
+     * @default undefined
+     * @example import fileUpload from "express-fileupload"
+     * @example uploader: fileUpload({ abortOnLimit: false, parseNested: true })
      * */
-    upload?: boolean | UploadOptions;
+    uploader?: RequestHandler;
     /**
-     * @desc Enable or configure response compression.
-     * @default false
+     * @desc Enable and configure compression
+     * @default undefined
+     * @example import compression from "compression"
+     * @example compressor: compression()
      */
-    compression?: boolean | CompressionOptions;
+    compressor?: RequestHandler;
     /**
      * @desc Enables parsing certain request payloads into raw Buffers (application/octet-stream by default)
      * @desc When enabled, use ez.raw() as input schema to get input.raw in Endpoint's handler