+++ tests/express-mock.ts
@@ -1,10 +1,6 @@
 // @see https://github.com/swc-project/jest/issues/14#issuecomment-970189585
 
 const expressJsonMock = jest.fn();
-const compressionMock = jest.fn();
-const fileUploadMock = jest.fn();
-jest.mock("compression", () => compressionMock);
-jest.mock("express-fileupload", () => fileUploadMock);
 
 const staticHandler = jest.fn();
 const staticMock = jest.fn(() => staticHandler);
@@ -26,12 +22,4 @@ appCreatorMock.static = staticMock;
 
 const expressMock = jest.mock("express", () => appCreatorMock);
 
-export {
-  compressionMock,
-  fileUploadMock,
-  expressMock,
-  appMock,
-  expressJsonMock,
-  staticMock,
-  staticHandler,
-};
+export { expressMock, appMock, expressJsonMock, staticMock, staticHandler };